+++ SolCARDIP/SolCARDIP.Librerias.EntidadesNegocio/beCarneIdentidadRelacionDependencia.cs
@@ -8,8 +8,8 @@ namespace SolCARDIP.Librerias.EntidadesNegocio
     public class beCarneIdentidadRelacionDependencia
     {
         public short TitularDependienteId { get; set; }                                   //CATD_STITULAR_DEPENDIENTE_ID
-        public short CarneIdentidadTitId { get; set; }                                    //CATD_SCARNE_IDENTIDAD_TIT_ID
-        public short CarneIdentidadDepId { get; set; }                                    //CATD_SCARNE_IDENTIDAD_DEP_ID
+        public int CarneIdentidadTitId { get; set; }                                    //CATD_SCARNE_IDENTIDAD_TIT_ID
+        public int CarneIdentidadDepId { get; set; }                                    //CATD_SCARNE_IDENTIDAD_DEP_ID
         public string Estado { get; set; }                                                //CATD_CESTADO
         public short UsuarioCreacion { get; set; }                                        //CATD_SUSUARIO_CREACION
         public string IpCreacion { get; set; }                                            //CATD_VIP_CREACION
